+++ srcs/Spark.Game/Entities/Character.cs
@@ -33,7 +33,7 @@ namespace Spark.Game.Entities
         public int MpPercentage { get; set; }
         public short MorphId { get; set; }
         public List<IBuff> Buffs { get; }
-        public byte Speed { get; set; }
+        public short Speed { get; set; }
         public Direction Direction { get; set; }
         public ICharacterInventory Inventory { get; }
         public IEnumerable<ISkill> Skills { get; set; }
