+++ srcs/Spark.Game/Entities/Npc.cs
@@ -30,7 +30,7 @@ namespace Spark.Game.Entities
         public int MpPercentage { get; set; }
         public short MorphId { get; set; }
         public List<IBuff> Buffs { get; }
-        public byte Speed { get; set; }
+        public short Speed { get; set; }
         public Direction Direction { get; set; }
 
         public bool Equals(IEntity other) => other != null && other.EntityType == EntityType && other.Id == Id;
