+++ srcs/Spark.Packet/Entities/Mv.cs
@@ -10,14 +10,14 @@ namespace Spark.Packet.Entities
         public EntityType EntityType { get; set; }
         public long EntityId { get; set; }
         public Vector2D Position { get; set; }
-        public byte Speed { get; set; }
+        public short Speed { get; set; }
 
         public void Construct(string[] packet)
         {
             EntityType = packet[0].ToEnum<EntityType>();
             EntityId = packet[1].ToLong();
             Position = new Vector2D(packet[2].ToShort(), packet[3].ToShort());
-            Speed = packet[4].ToByte();
+            Speed = packet[4].ToShort();
         }
     }
 }
